# internal/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API request bodies and responses are usually JSON-encoded, except for binary data and some of the OAuth2 endpoints. Note that the `Content-Type` header must be set to `application/json` for this, as the request may otherwise fail with error 400.\nWhen clients have a valid access token, e.g. obtained through the `POST /api/v1/session` or `POST /api/v1/oauth/token` endpoint, they can use a standard Bearer Authorization header to authenticate their requests. Submitting the access token with a custom `X-Auth-Token` header is supported as well.",
        "title": "PhotoPrism API",
        "contact": {},
        "version": "v1"
    },
    "host": "demo.photoprism.app",
    "paths": {
        "/api/v1/albums": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Albums"
                ],
                "summary": "finds albums and returns them as JSON",
                "operationId": "SearchAlbums",
                "parameters": [
                    {
                        "maximum": 100000,
                        "minimum": 1,
                        "type": "integer",
                        "description": "maximum number of results",
                        "name": "count",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100000,
                        "minimum": 0,
                        "type": "integer",
                        "description": "search result offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "favorites",
                            "name",
                            "title",
                            "added",
                            "edited"
                        ],
                        "type": "string",
                        "description": "sort order",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/search.Album"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Albums"
                ],
                "summary": "creates a new album",
                "operationId": "CreateAlbum",
                "parameters": [
                    {
                        "description": "properties of the album to be created (currently supports Title and Favorite)",
                        "name": "album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.Album"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Album"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Album": {
            "type": "object",
            "properties": {
                "Caption": {
                    "type": "string"
                },
                "Category": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "CreatedBy": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "Filter": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "Location": {
                    "type": "string"
                },
                "Month": {
                    "type": "integer"
                },
                "Notes": {
                    "type": "string"
                },
                "Order": {
                    "type": "string"
                },
                "ParentUID": {
                    "type": "string"
                },
                "Path": {
                    "type": "string"
                },
                "Private": {
                    "type": "boolean"
                },
                "PublishedAt": {
                    "type": "string"
                },
                "Slug": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                },
                "Template": {
                    "type": "string"
                },
                "Thumb": {
                    "type": "string"
                },
                "ThumbSrc": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "Year": {
                    "type": "integer"
                }
            }
        },
...
        "form.Album": {
            "type": "object",
            "properties": {
                "Caption": {
                    "type": "string"
                },
                "Category": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "Filter": {
                    "type": "string"
                },
                "Location": {
                    "type": "string"
                },
                "Notes": {
                    "type": "string"
                },
                "Order": {
                    "type": "string"
                },
                "Private": {
                    "type": "boolean"
                },
                "Template": {
                    "type": "string"
                },
                "Thumb": {
                    "type": "string"
                },
                "ThumbSrc": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                }
            }
        },
...
        "i18n.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "search.Album": {
            "type": "object",
            "properties": {
                "Caption": {
                    "type": "string"
                },
                "Category": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "Day": {
                    "type": "integer"
                },
                "DeletedAt": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Favorite": {
                    "type": "boolean"
                },
                "Filter": {
                    "type": "string"
                },
                "LinkCount": {
                    "type": "integer"
                },
                "Location": {
                    "type": "string"
                },
                "Month": {
                    "type": "integer"
                },
                "Notes": {
                    "type": "string"
                },
                "Order": {
                    "type": "string"
                },
                "ParentUID": {
                    "type": "string"
                },
                "Path": {
                    "type": "string"
                },
                "PhotoCount": {
                    "type": "integer"
                },
                "Private": {
                    "type": "boolean"
                },
                "Slug": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                },
                "Template": {
                    "type": "string"
                },
                "Thumb": {
                    "type": "string"
                },
                "ThumbSrc": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "Year": {
                    "type": "integer"
                }
            }
        },
...
    },
...
}